Control de Errores

+++ ProductoMVVMSQLite-master/ProductoMVVMSQLite/ViewModels/ProductoViewModel.cs
@@ -16,10 +16,11 @@ namespace ProductoMVVMSQLite.ViewModels
     public class ProductoViewModel
     {
         public ObservableCollection<Producto> ListaProductos { get; set; }
-        public Producto _producto;
-        public ProductoViewModel() {
 
-            Util.ListaProductos = new ObservableCollection<Producto>(App.productoRepository.GetAll()); ;
+        public Producto ProductoSeleccionado {get;set;}
+
+        public ProductoViewModel() {
+            Util.ListaProductos = new ObservableCollection<Producto>(App.productoRepository.GetAll());
 
             ListaProductos = Util.ListaProductos;
         
@@ -30,24 +31,58 @@ namespace ProductoMVVMSQLite.ViewModels
             {
                await App.Current.MainPage.Navigation.PushAsync(new NuevoProductoPage());
             });
-       
-        public ICommand EliminarProducto =>
-        new Command<Producto>((producto) =>
-        {
-           
-            App.productoRepository.Delete(producto);
-            ListaProductos.Remove(producto);
-            Util.ListaProductos.Clear();
-            App.productoRepository.GetAll().ForEach(Util.ListaProductos.Add);
-        });
+
         public ICommand EditarProducto =>
-        new Command<Producto>(async (producto) =>
-        {
-           
-            await App.Current.MainPage.Navigation.PushAsync(new DetailsPage(producto));
-        });
+            new Command(async () =>
+            {
+                if (ProductoSeleccionado != null)
+                {
+                    int IdProducto = ProductoSeleccionado.IdProducto;
+                    //ProductoSeleccionado = null;
+                    await App.Current.MainPage.Navigation.PushAsync(new NuevoProductoPage(IdProducto));
+                    ProductoSeleccionado = null;
+                }
+              
+
+            });
+        public ICommand ModificarDesdeBotonCommand =>
+            new Command<Producto>(async (producto) =>
+            {
+                if (producto != null)
+                {
+                    await App.Current.MainPage.Navigation.PushAsync(new NuevoProductoPage(producto.IdProducto));
+                }
+            });
+
+        public ICommand EliminarDesdeBotonCommand =>
+            new Command<Producto>(async (producto) =>
+            {
+                if (producto != null)
+                {
+                    bool confirmarEliminar = await App.Current.MainPage.DisplayAlert("Eliminar", $"¿Eliminar {producto.Nombre}?", "Sí", "No");
+
+                    if (confirmarEliminar)
+                    {
+                        App.productoRepository.Delete(producto);
+                        Util.ListaProductos.Remove(producto);
+                    }
+                }
+            });
+
+        public ICommand VerProductoDesdeBotonCommand =>
+            new Command<Producto>(async (producto) =>
+            {
+                if (producto != null)
+                {
+                    var detalleProductoViewModel = new DetalleProductoViewModel(producto);
+                    var detalleProductoPage = new DetalleProductoPage { BindingContext = detalleProductoViewModel };
+                    await App.Current.MainPage.Navigation.PushAsync(detalleProductoPage);
+                }
+            });
 
 
-        
+
+
+
     }
 }
